refactor: rename command packages for better organization
